Refactored test structure to match new module structure
