+++ test/objects/longview_test.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from test.base import ClientBaseCase
 from linode_api.objects.base import MappedObject
 
-from linode_api.objects import LongviewClient
+from linode_api.objects import LongviewClient, LongviewSubscription
 
 
 class LongviewClientTest(ClientBaseCase):
@@ -54,3 +54,24 @@ class LongviewClientTest(ClientBaseCase):
             client.delete()
 
             self.assertEqual(m.call_url, '/longview/clients/1234')
+
+
+class LongviewSubscriptionTest(ClientBaseCase):
+    """
+    Tests methods of the LongviewSubscription class
+    """
+    def test_get_subscription(self):
+        """
+        Tests that a subscription is loaded correctly by ID
+        """
+        sub = LongviewSubscription(self.client, "longview-40")
+        self.assertEqual(sub._populated, False)
+
+        self.assertEqual(sub.label, 'Longview Pro 40 pack')
+        self.assertEqual(sub._populated, True)
+
+        self.assertEqual(sub.clients_included, 40)
+
+        self.assertIsInstance(sub.price, MappedObject)
+        self.assertEqual(sub.price.hourly, .15)
+        self.assertEqual(sub.price.monthly, 100)
